+++ examples/download/src/download_error.rs
@@ -1,4 +1,10 @@
+use std::path::PathBuf;
+
+#[cfg(feature = "error_reporting")]
+use peace::miette::{self, SourceSpan};
+
 /// Error while managing a file download.
+#[cfg_attr(feature = "error_reporting", derive(peace::miette::Diagnostic))]
 #[derive(Debug, thiserror::Error)]
 pub enum DownloadError {
     #[error("Failed to open destination file.")]
@@ -7,13 +13,38 @@ pub enum DownloadError {
     DestMetadataRead(#[source] std::io::Error),
     #[error("Failed to read destination file contents.")]
     DestFileRead(#[source] std::io::Error),
-    #[error("Failed to open destination file for writing.")]
-    DestFileCreate(#[source] std::io::Error),
+
+    #[error("Failed to open `{}` for writing.", dest.display())]
+    #[cfg_attr(
+        feature = "error_reporting",
+        diagnostic(
+            code(download::dest_file_create),
+            help(
+                "Ensure that `{}` is not a directory, or rerun `download init` with a different file path.",
+                dest.display()))
+    )]
+    DestFileCreate {
+        /// Approximation of the init command that defined the destination path.
+        #[cfg_attr(feature = "error_reporting", source_code)]
+        init_command_approx: String,
+        #[cfg(feature = "error_reporting")]
+        #[label = "defined here"]
+        dest_span: SourceSpan,
+        /// Destination file path.
+        dest: PathBuf,
+        /// Underlying IO error
+        #[source]
+        error: std::io::Error,
+    },
     #[error("Failed to delete destination file.")]
     DestFileRemove(#[source] std::io::Error),
     #[error("Failed to parse source URL.")]
     SrcUrlParse(url::ParseError),
-    #[error("Failed to parse source URL.")]
+    #[cfg_attr(
+        feature = "error_reporting",
+        diagnostic(code(download::src_get), help("Check that the URL is reachable."))
+    )]
+    #[error("Failed to fetch from URL.")]
     SrcGet(#[source] reqwest::Error),
     #[error("Failed to fetch source file metadata. Response status code: {status_code}")]
     SrcFileUndetermined { status_code: reqwest::StatusCode },
@@ -27,11 +58,26 @@ pub enum DownloadError {
     // Framework errors
     /// A `peace` runtime error occurred.
     #[error("A `peace` runtime error occurred.")]
-    PeaceRtError(#[from] peace::rt_model::Error),
+    PeaceRtError(
+        #[cfg_attr(feature = "error_reporting", diagnostic_source)]
+        #[source]
+        #[from]
+        peace::rt_model::Error,
+    ),
 
     // Scaffolding errors
     #[error("Failed to initialize tokio runtime.")]
     TokioRuntimeInit(#[source] std::io::Error),
+    #[cfg(not(target_arch = "wasm32"))]
+    #[error("Failed to read current executable path.")]
+    CurrentExeRead(#[source] std::io::Error),
+    #[cfg(not(target_arch = "wasm32"))]
+    #[error("Failed to get current executable name from path.")]
+    CurrentExeNameRead,
+    /// This one should be relatively unreachable.
+    #[cfg(not(target_arch = "wasm32"))]
+    #[error("Failed to format string in memory.")]
+    FormatString(#[source] std::fmt::Error),
 
     // WASM errors.
     #[cfg(target_arch = "wasm32")]