+++ examples/download/src/main.rs
@@ -9,7 +9,32 @@ use download::{
     workspace_and_graph_setup, DownloadArgs, DownloadCommand, DownloadError, DownloadProfileInit,
 };
 
+#[cfg(not(feature = "error_reporting"))]
 pub fn main() -> Result<(), DownloadError> {
+    run()
+}
+
+#[cfg(feature = "error_reporting")]
+pub fn main() -> peace::miette::Result<(), peace::miette::Report> {
+    // Important to return `peace::miette::Report` instead of calling
+    // `IntoDiagnostic::intoDiagnostic` on the `Error`, as that does not present the
+    // diagnostic contextual information to the user.
+    //
+    // See <https://docs.rs/miette/latest/miette/trait.IntoDiagnostic.html#warning>.
+
+    // The explicit mapping for `PeaceRtError` appears to be necessary to display
+    // the diagnostic information. i.e. `miette` does not automatically delegate to
+    // the #[diagnostic_source].
+    //
+    // This is fixed by <https://github.com/zkat/miette/pull/170>.
+
+    run().map_err(|download_error| match download_error {
+        DownloadError::PeaceRtError(err) => peace::miette::Report::from(err),
+        other => peace::miette::Report::from(other),
+    })
+}
+
+pub fn run() -> Result<(), DownloadError> {
     let runtime = tokio::runtime::Builder::new_current_thread()
         .thread_name("main")
         .thread_stack_size(3 * 1024 * 1024)